+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- ZookeeperZnode.spec.clusterRef.namespace now defaults to .metadata.namespace ([#382]).
+
 ### Changed
 
 - Shut down gracefully ([#338]).
@@ -14,6 +18,7 @@ All notable changes to this project will be documented in this file.
 [#338]: https://github.com/stackabletech/zookeeper-operator/pull/338
 [#340]: https://github.com/stackabletech/zookeeper-operator/pull/340
 [#352]: https://github.com/stackabletech/zookeeper-operator/pull/352
+[#382]: https://github.com/stackabletech/zookeeper-operator/pull/382
 [#384]: https://github.com/stackabletech/zookeeper-operator/pull/384
 
 ## [0.8.0] - 2021-12-22